Refactor Sleep Diary Entries
